feat: integrate sign-in and sign-up buttons (#4)
