+++ app/(auth)/sign-in/components/sign-in-form.tsx
@@ -1,0 +1,118 @@
+"use client";
+
+import { useForm } from "react-hook-form";
+import * as z from "zod";
+import { zodResolver } from "@hookform/resolvers/zod";
+import Link from "next/link";
+
+import { Input } from "../../../../components/ui/input";
+import { Button } from "../../../../components/ui/button";
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "../../../../components/ui/form";
+import { useRouter, useSearchParams } from "next/navigation";
+import { Loader2 } from "lucide-react";
+import { useToast } from "../../../../components/ui/use-toast";
+import { useAuth } from "@/lib/auth";
+import { useState } from "react";
+import { Separator } from "@/components/ui/separator";
+
+export const signInSchema = z.object({
+  username: z.string().min(3).max(20),
+  password: z
+    .string()
+    .regex(
+      new RegExp(/^(?=.*[a-zA-Z])(?=.*[!@#$%^*+=-])(?=.*[0-9]).{8,15}$/),
+      "Password must contain at least 8 characters, including letters, numbers, and special characters."
+    ),
+});
+
+export default function SignInForm() {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const { signIn } = useAuth();
+  const { toast } = useToast();
+
+  const [loading, setLoading] = useState(false);
+
+  const form = useForm<z.infer<typeof signInSchema>>({
+    resolver: zodResolver(signInSchema),
+    defaultValues: {
+      username: "",
+      password: "",
+    },
+  });
+
+  const onSubmit = async (values: z.infer<typeof signInSchema>) => {
+    setLoading(true);
+    const dst = searchParams.get("redirect") || "/channels";
+    const ok = await signIn(values);
+
+    if (ok) {
+      router.replace(dst);
+    } else {
+      toast({
+        title: "Your sign in request is failed.",
+        description: "Invalid ID or password.",
+        variant: "destructive",
+      });
+    }
+
+    setLoading(false);
+  };
+
+  return (
+    <div className="flex flex-col gap-y-4">
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
+          <FormField
+            control={form.control}
+            name="username"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Username</FormLabel>
+                <FormControl>
+                  <Input placeholder="username" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="password"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Password</FormLabel>
+                <FormControl>
+                  <Input placeholder="password" type="password" {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <Button
+            type="submit"
+            className="bg-kookmin dark:text-white w-full"
+            disabled={loading}
+          >
+            {loading ? (
+              <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+            ) : (
+              "Submit"
+            )}
+          </Button>
+        </form>
+      </Form>
+      <Button className="w-full" asChild>
+        <Link href="/sign-up">Don't have an account?&nbsp;<strong>Sign Up</strong></Link>
+      </Button>
+    </div>
+  );
+}
