+++ app/(auth)/sign-in/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import SignInForm from "@/components/auth/sign-in-form";
+import SignInForm from "@/app/(auth)/sign-in/components/sign-in-form";
 import Loading from "@/components/loading";
 import { useAuth } from "@/lib/auth";
 import { useRouter } from "next/navigation";
