+++ app/components/auth-provider.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import { AuthContext, SignInParams, SignUpParams } from "@/lib/auth";
 import type { User } from "@/types";
-import { toast } from "./ui/use-toast";
+import { toast } from "../../components/ui/use-toast";
 
 export default function AuthProvider({
   children,
