+++ app/components/create-channel-button.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 
-import { Button } from "./ui/button";
+import { Button } from "../../components/ui/button";
 
 export default function CreateChannelButton() {
   return (
