--- components/auth/sign-up-button.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import { Button } from "../ui/button";
-import Link from "next/link";
-
-export default function SignUpButton() {
-  return (
-    <Button asChild className="bg-kookmin dark:text-white">
-      <Link href="/sign-up">Sign Up</Link>
-    </Button>
-  );
-}
