--- components/auth/sign-up-form.tsx
@@ -1,139 +1,0 @@
-"use client";
-
-import { useForm } from "react-hook-form";
-import * as z from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-
-import { Input } from "../ui/input";
-import { Button } from "../ui/button";
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "../ui/form";
-import { useRouter } from "next/navigation";
-import React from "react";
-import { Loader2 } from "lucide-react";
-import { useToast } from "../ui/use-toast";
-import { useAuth } from "@/lib/auth";
-
-export const signUpSchema = z
-  .object({
-    username: z.string().min(3).max(20),
-    password: z
-      .string()
-      .regex(
-        new RegExp(/^(?=.*[a-zA-Z])(?=.*[!@#$%^*+=-])(?=.*[0-9]).{8,15}$/),
-        "Password must contain at least 8 characters, including letters, numbers, and special characters."
-      ),
-    confirmPassword: z.string(),
-  })
-  .refine((data) => data.password === data.confirmPassword, {
-    message: "Passwords don't match",
-    path: ["confirmPassword"],
-  });
-
-export default function SignUpForm() {
-  const [loading, setLoading] = React.useState(false);
-  const router = useRouter();
-  const { signUp } = useAuth();
-  const { toast } = useToast();
-
-  const form = useForm<z.infer<typeof signUpSchema>>({
-    resolver: zodResolver(signUpSchema),
-    defaultValues: {
-      username: "",
-      password: "",
-      confirmPassword: "",
-    },
-  });
-
-  const onSubmit = async (values: z.infer<typeof signUpSchema>) => {
-    setLoading(true);
-    const ok = await signUp({
-      username: values.username,
-      password: values.password,
-    });
-
-    if (ok) {
-      toast({
-        title: "Your sign up request is successful.",
-        description: "You can sign in now.",
-        variant: "success",
-      });
-      router.push("/sign-in");
-    } else {
-      toast({
-        title: "Your sign up request is failed.",
-        description: "The id already exists.",
-        variant: "destructive",
-      });
-    }
-    setLoading(false);
-  };
-
-  return (
-    <Form {...form}>
-      <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-        <FormField
-          control={form.control}
-          name="username"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Username</FormLabel>
-              <FormControl>
-                <Input placeholder="username" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="password"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Password</FormLabel>
-              <FormControl>
-                <Input placeholder="password" type="password" {...field} />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="confirmPassword"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>Confirm Password</FormLabel>
-              <FormControl>
-                <Input
-                  placeholder="confirm password"
-                  type="password"
-                  {...field}
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <Button
-          type="submit"
-          className="bg-kookmin dark:text-white"
-          disabled={loading}
-        >
-          {loading ? (
-            <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-          ) : (
-            "Submit"
-          )}
-        </Button>
-      </form>
-    </Form>
-  );
-}
