+++ components/channels/video-player.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useEffect, useRef } from "react";
-import { useSocket } from "../socket-provider";
+import { useSocket } from "../../app/components/socket-provider";
 
 /**
  * @param broadcaster broadcaster's username
